feat: tidy up and some renaming
