+++ Microservices/Profile/src/main/java/cs4337/group_8/ProfileService/TemplateProjectApplication.java
@@ -1,4 +1,4 @@
-package cs4337.group_8.TemplateProject;
+package cs4337.group_8.ProfileService;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
