+++ Microservices/Profile/src/test/java/cs4337/group_8/TemplateProject/TemplateProjectApplicationTests.java
@@ -1,4 +1,4 @@
-package cs4337.group_8.TemplateProject;
+package cs4337.group_8.ProfileService;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
